migrate MapBox
