+++ superset/assets/src/visualizations/MapBox/adaptor.jsx
@@ -1,0 +1,5 @@
+import createAdaptor from '../../utils/createAdaptor';
+import Component from './MapBox';
+import transformProps from './transformProps';
+
+export default createAdaptor(Component, transformProps);
